requirements.txt
langchain==0.3.19
langchain_community==0.3.18
langchain_google_genai==2.0.11
protobuf==5.29.3
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.42.2
streamlit_chat==0.1.1
faiss-cpu